requirements.txt
Flask
Flask-SQLAlchemy
pymysql
requests
Werkzeug
